Made two changes to maketea, which are reflected in the generated code:
- Every AST class now has a visit(AST_visitor*) and 
  transform_children(AST_transform*) method, which call the appropriate 
  method in the visitor/transform. It is convenient to add these methods 
  (means the programmer doesn't have to know which method in the visitor to 
  call, which isn't always trivial) and since both these methods return void,
  it doesn't cause problems with novariance. 

- maketea generates a mapping from class names to integers (the class IDs), 
  but this mapping was not explicitly available once the code was generated.
  Instead, we now add a static const member to every class which corresponds
  to its identifier, and use those members as identifiers; so, for instance:

  if(node->classid() == AST_eval_expr::ID) ..

Modified the lexer to allow for more opening and starting tags.

Any calls to "php_script->transform" have been changed into calls to
"php_script->transform_children"; this is more accurate, because the transform
wasn't allowed to modify the script itself (if it did, we never used the 
transformed script). This has one consequence, which is that the transform
should not use pre_php_script to do anything; this was the case in only once
transform (Process_includes), which now overrides children_php_script instead.