An unforutnately large commit.
Added a warning about break levels, if the expression part of a break or continue statement is provided (or NULL), and is greater than the number of levels it is possible to jump. This causes a lot of test cases to throw errors. These were fixed either by making the expression not immediately obvious, or by moving the test to errors.

So homogenize_X functions were updated for this.

There is a case where the warnings only appeared due to include () statements. My simplistic method of regexing for the error no longer worked as a result. The solution is to use phc to parse the file, however, that will lead to the same errors that we want to check. As a result, I added the --dont-fail command line option, with which phc_error messages dont exit (-1). It seemed that the current use cases of this are tools and applications which need the ast, much like this problem, and that a simple hack like this is probably the better solution. Of course, it can lead to segfaults in some cases, and phc will still fail if the program doesnt parse. This is desirable/acceptable, however.

If phc fails with the --dont-fail option, then parse the file the old way for testing.


Because I'm committing phc.cpp, it makes sense to commit the other changes I had made to it, which adds another run of the include pass at a later stage. The boolean is to indicate whether or not to use gotos in the lowered code. It also removes an unused function.