Fix bug0025, from Gary Wassermann.
I'm not sure about this, truth be told. It moves the post-op shredding into early_shredder, so that class variables dont get shredded before the post_op does. It would be better if we had a definition of shredding priority, but this is perhaps left to a later date

I remember a test along the lines of
	a()++;
which broke with shredding. Of course, that code isnt legal, but I cant find the test now. If that was legal, this change would probably break it. But all tests pass, so we should be OK.