The second step is moving to using a temp array.
Some functions, such as write_array, need to separate the zval. In order to do this, they must be able to access the symbol table containing the zval. We handled this by passing the symbol table, name, length and hashval required to find the original. Now, when we look up a value, using get_st_entry, we add EG(uninitialized_zval_ptr) to the symbol table, and return a pointer into it. It is then possible to separate using the returned pointer to zval, since it points into the symbol table.

Tests all pass, and one of the refcount tests spontaneously started passing, so it seems this is a more correct approach.