Fix Lower_control_flow of break statements.
The major problem was that the shredder had been removed. Instead, we use eval_var and push_back_pieces to get the same effect. The shredder was too much hassle for this one file.

If the break level is 0 (ie we are in the outer loop), then always print an error, regardless of the parameter.

At any depth, if we know the parameter to the break, we can jump directly to its label, instead of looping through possible versions (we change to a vector for this).