This adds superglobals other than GLOBALS. It also adds an optimization to remove those globals.
We look through the variables used in a method, and if there is no reflection
or run-time code generation, then the globals can be removed if they are
unused.

Additionally, we make Lift_functions_and_classes non-optional.

Overall, this results in a 4% speed increase, but this is really about 15% increase in each of the recursive functions, which benefit from having GLOBALS removed.

Note that this breaks a number of tests, specifically anything with nested functions. Rather than fix this, I'm going to replace nested functions with MIR intrinsics, which we talked about a long time ago. I'm committing this now since I need the results.