Each unparser uses other unparsers to handle its foreign nodes, or its own nodes after they are folded into AST. However, none of the parser's really shared state, and in the ones which did (generally using the AST_unparser as the base) did it badly. In particular, most IRs had another IR to which some tasks were delegated, with which they did not share state.
This removes the state from the PHP_unparser, and enables it to be shared. Unparsers which delegate soem responsabilities to other unparsers ensure that the state is passed along. This fixes nearly all unparsing bugs, and solves a few problems which did not manifest in tests (these being that sometimes <?php appeared in the unparsed output, and that lines would frequently not be indented).

The only remaining bug is that cb_HIR-to-MIr does not work, but this test will be changed to check for the same output as mir_dump.