The XML test actually ran each pass twice. "phc --read-xml=lfc --xdump=clar" reads the IR at the start of the lfc pass, runs the lfc pass, runs the clar pass, then prints out XML. The next step is to run "phc --read-xml=clar --xdump=...", which runs the clar pass for the second time. This means that moving main into the __MAIN__ method will happen twice, which is why all the passes broke.
To fix it, each step of the xml test is in the form "phc --read-xml=pass --xdump=pass", which fixes the problem. The only new problem is that "--xdump=AST-to-HIR" dumps AST, which "--read-xml=hir" can't read. In this case, we do "--read-xml=AST-to-HIR --xdump=hir" (and the same for HIR-to-MIR).