The previous fix for froreach was insufficient. Given:
foreach ($arr as $key => &$val)
{
	...
}

it generated:

...
$val =& $iter->current ();

which does not actually return a reference (it isnt possible to return a reference from a function which isnt declared to return-by-ref). Instead, we must use a slightly different version of the old form:

$val =& $arr[$iter->key()];


This is local to uppering, doesn't break optimizations, and allows the IR clean-up.